pixelEditor: DSCO CustomDialog for the modal chrome; extract toolDefinitions
CustomDialog replaces the hand-rolled DialogChrome (same focus-trap/Escape
hooks underneath) and adds scroll lock, a close button, and the dialog
description; the editor's panel styles win via a display: contents host's
parent selector, not stylesheet order. The decode-phase backdrop and the
custom z-index are gone: decode is fast enough to render nothing until the
dialog arrives in final form, and DSCO's modal-backdrop layer (1040) already
clears the Lab2 chrome (tab bar 5, guide 71). The TOOLS array moves to
toolDefinitions.tsx.
